Factor out lifters
