+++ CMakeLists.txt
@@ -2,13 +2,13 @@ cmake_minimum_required(VERSION 3.1)
 
 set(CMAKE_POSITION_INDEPENDENT_CODE ON)
 set(CMAKE_CXX_STANDARD 11)
-set(CMAKE_CXX_FLAGS "-Werror -Wall -Wextra -Wno-parentheses")
+set(CMAKE_CXX_FLAGS "-Werror -Wall -Wextra -Wno-parentheses -Wno-unused-function")
 set(CMAKE_EXPORT_COMPILE_COMMANDS ON)
 set(CMAKE_CXX_STANDARD 17)
 
 find_library(BINJA_CORE_LIBRARY
-    NAMES binaryninjacore libbinaryninjacore.so.1
-    PATHS /opt/binaryninja)
+        NAMES binaryninjacore libbinaryninjacore.so.1
+        PATHS /opt/binaryninja)
 
 add_subdirectory(binaryninja-api)
 
@@ -33,5 +33,5 @@ include_directories(capstone/include)
 
 add_subdirectory(capstone)
 
-add_library(aarch64_extension SHARED aarch64_extension.cpp)
+add_library(aarch64_extension SHARED A64.cpp A64.h LiftDT.cpp LiftDT.h LiftFP.cpp LiftFP.h LiftGR.cpp LiftGR.h LiftVT.cpp LiftVT.h)
 target_link_libraries(aarch64_extension binaryninjaapi capstone-static)
